+++ composeApp/src/commonMain/kotlin/com/andef/myfinance/feature/income_common/income_analysis/presentation/IncomeAnalysisIntent.kt
@@ -9,7 +9,7 @@ sealed class IncomeAnalysisIntent {
     ) : IncomeAnalysisIntent()
 
     data class GetIncomesForPdf(
-        val onSuccess: (List<Pair<LocalDate, Double>>, maxDate: LocalDate, minDate: LocalDate) -> Unit,
+        val onSuccess: (List<Pair<LocalDate, Long>>, maxDate: LocalDate, minDate: LocalDate) -> Unit,
         val onError: (String) -> Unit,
         val startDate: LocalDate,
         val endDate: LocalDate
